requirements.txt
asn1crypto==0.24.0
Authlib==0.10
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
Flask==1.0.2
geographiclib==1.49
geopy==1.20.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2==2.7.7
pycparser==2.19
python-dotenv==0.10.1
requests==2.21.0
six==1.12.0
urllib3==1.24.2
Werkzeug==0.15.1
